# openAPI.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Marketplace Datasink application",
    "description": "Datasink provides a place to store your simulation data and retrieve it back. The data can be both binary data and graph data.",
    "contact": {
      "name": "Kiran Kumaraswamy",
      "url": "https://materials-marketplace.eu/",
      "email": "kiran.kumaraswamy@iwm.fraunhofer.de"
    },
    "version": "0.1.0",
    "x-api-version": "0.3.0",
    "x-products": [
          {
              "name": "product A",
              "productId": null
          }
      ]
  },
  "servers": [
    {
      "url": "https://data-sink-app.materials-data.space/"
    }
  ],
...
  "paths": {
...
    "/heartbeat": {
      "get": {
        "summary": "Check if app is alive",
        "operationId": "heartbeat",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}